+++ pages/index.js
@@ -1,0 +1,58 @@
+import Head from 'next/head'
+import Image from 'next/image'
+import { Inter } from 'next/font/google'
+import styles from '@/styles/Home.module.css'
+
+const inter = Inter({ subsets: ['latin'] })
+
+export default function Home() {
+  return (
+    <>
+      <Head>
+        <title>Create Next App</title>
+        <meta name="description" content="Generated by create next app" />
+        <meta name="viewport" content="width=device-width, initial-scale=1" />
+        <link rel="icon" href="/favicon.ico" />
+
+        <script src='/try.js    '></script>
+      </Head>
+      <nav className={styles.mainnav}>
+        <ul>
+          <li>Home</li>
+          <li>Blog</li>
+          <li>About</li>
+          <li>Contact</li>
+        </ul>
+      </nav>
+      <main className={`${styles.main} ${inter.className}`}>
+        {/* <div className={styles.description}> */}
+          <h1 className='hding'>
+            Hunting Coders
+            </h1>
+        
+          
+        {/* </div> */}
+
+        <div className={styles.center}>
+         <h2>
+          Populer Blog
+         </h2>
+
+        </div>
+
+        <div className="blog">
+          <div className="blogImage">
+            <h3>the blog content
+            </h3>
+            <h3>the blog content
+            </h3>
+            <h3>the blog content
+            </h3>
+            <h3>the blog content
+            </h3>
+          </div>
+        </div>
+      </main>
+    </>
+  )
+}
