+++ public/try.js
@@ -1,0 +1,1 @@
+alert("jfbjbjb")
