+++ styles/Home.module.css
@@ -3,8 +3,11 @@
   flex-direction: column;
   justify-content: space-between;
   align-items: center;
-  padding: 6rem;
+  padding: 2rem;
   min-height: 100vh;
+
+
+
 }
 
 .description {
@@ -226,4 +229,28 @@
   to {
     transform: rotate(0deg);
   }
+}
+
+
+
+
+.mainnav ul
+{
+  display: flex;
+  /* text-decoration: none; */
+  align-items: center;
+  justify-content: center;
+  list-style: none;
+margin: 50px;  
+}
+.mainnav ul li{
+list-style: none;
+margin: 30px;
+display: flex;
+flex-wrap: wrap;
+}
+.hding{
+  margin: 10px;
+  align-items: center;
+  font-family: cursive;
 }
